feat: Move `IfExpr` to `spark-expr` crate (#653)
