+++ native/core/Cargo.toml
@@ -77,7 +77,8 @@ once_cell = "1.18.0"
 regex = "1.9.6"
 crc32fast = "1.3.2"
 simd-adler32 = "0.3.7"
-datafusion-comet-spark-expr = { path = "../spark-expr", version = "0.1.0" }
+datafusion-comet-spark-expr = { workspace = true }
+datafusion-comet-utils = { workspace = true }
 
 [build-dependencies]
 prost-build = "0.9.0"
