+++ native/utils/Cargo.toml
@@ -16,8 +16,8 @@
 # under the License.
 
 [package]
-name = "datafusion-comet-spark-expr"
-description = "DataFusion expressions that emulate Apache Spark's behavior"
+name = "datafusion-comet-utils"
+description = "DataFusion Comet Utilities"
 version = { workspace = true }
 homepage = { workspace = true }
 repository = { workspace = true }
@@ -27,14 +27,8 @@ license = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
-arrow = { workspace = true }
-arrow-schema = { workspace = true }
-datafusion = { workspace = true }
-datafusion-common = { workspace = true }
-datafusion-functions = { workspace = true }
-datafusion-physical-expr = { workspace = true }
-datafusion-comet-utils = { workspace = true }
+datafusion-physical-plan = { workspace = true }
 
 [lib]
-name = "datafusion_comet_spark_expr"
+name = "datafusion_comet_utils"
 path = "src/lib.rs"
